tracking_service/requirements.txt
redis
opencv-python
pyyaml
numpy
ultralytics
supervision
lap
lapx
Cython
db-sqlite3
shapely
laptrack
torchreid
tensorboard
filterpy